move mock data to fixtures folder
